Unslash our input/data
WordPress still escapes all $_GET and $_POST input as a way to maintain
compatibility with the ages old default of magic_quotes being on in very
old PHP versions. For all of our input (including data we saved to the DB
in the past) we're using wp_unslash to catch everything and save/display
it cleanly.

See: https://core.trac.wordpress.org/ticket/18322